moved privtangle and nodeconn tests to cluster tsts
